+++ src/src/App.vue
@@ -1,132 +1,25 @@
 <script setup lang="ts">
-import { invoke, dialog } from '@tauri-apps/api'
-import { onMounted, ref } from 'vue';
-import { listen } from '@tauri-apps/api/event'
-
-let hasHover = ref(false);
-let dragTarget: any;
-let filePath = ref("");
-let hash_result_512 = ref("");
-let hash_result_256 = ref("");
-
-type HashResult = {
-  blake512: string,
-  blake256: string
-}
-
-onMounted(() => {
-  invoke('greet', { name: 'World' })
-    // `invoke` returns a Promise
-    .then((response) => console.log(response))
+import { PropType } from 'vue';
+import FileStatus from './components/FileStatus.vue';
+import { HashResult } from './types/HashResult'
+const props = defineProps({
+  hashResult: Object as PropType<HashResult>
 })
 
-listen('tauri://file-drop', event => {
-  console.log(event);
-  hasHover.value = false;
-  dropFile(event);
-})
-
-listen('tauri://file-drop-cancelled', event => {
-  console.log(event);
-  hasHover.value = false;
-})
-
-/**
- * Listens to the dropenter event and sets the event target
- * Applies animation
- * @param event DropEnter Event
- */
-function startDragAnimation(event: any) {
-  dragTarget = event.target;
-  event.stopPropagation();
-  event.preventDefault();
-  console.log("Start Drag animation", event);
-  hasHover.value = true;
-}
-
-function openFileDialog() {
-  dialog.open({ title: "Please select a file...", multiple: false }).then((selectedFiles) => {
-    if (selectedFiles != null) {
-      if (!Array.isArray(selectedFiles)) {
-        filePath.value = selectedFiles
-        getFileInfo(filePath.value);
-      }
-    }
-  });
-}
-
-/**
- * Listens to the dropleave event and checks the event target
- * Removes animation
- * @param event DragLeave Event
- */
-function stopDragAnimation(event: any) {
-  if (dragTarget === event.target) {
-    hasHover.value = false;
-    event.preventDefault();
-    dragTarget = null;
-    console.log("Stop Drag animation", event);
-    event.stopPropagation();
-    event.preventDefault();
-  }
-}
-
-function getFileInfo(path: string) {
-  console.log("Check for file path", path);
-
-  invoke('check_file', { filePath: path }).then((result: any) => {
-    let res = result as HashResult;
-    console.log("Result", result);
-    hash_result_256.value = res.blake256;
-    hash_result_512.value = res.blake512;
-  }).catch((error) => {
-    console.log(error);
-  })
-}
-
-function dropFile(event: any) {
-  filePath.value = event.payload[0];
-  getFileInfo(filePath.value);
-}
 </script>
 
 <template>
-  <div class="p-2 m-2">
-    <div @dragenter="startDragAnimation($event)" @dragleave="stopDragAnimation($event)"
-      @dragexit="stopDragAnimation($event)">
-      <div class="group h-28 bg-gray-50 rounded-md border-2 border-gray-500 border-dashed">
-        <div class="pt-4 flex place-content-center">
-          <img v-bind:class="{ 'animate-bounce': hasHover }" class="h-12 w-12" src="./assets/folder.png" alt="Folder">
-        </div>
-        <div class="flex place-content-center text-xl font-medium text-black"> Drop File Here</div>
-      </div>
+  <FileStatus  />
+  <div style="height: 250px;">
+    <label class="block text-left text-gray-600 font-bold mb-1 pt-2" for="blake2b">Blake2s256</label>
+    <div
+      class="bg-slate-700 break-words text-white max-h-56 overflow-auto fullshadow-md rounded px-2 pt-1 pb-1  text-sm">
+      {{ hashResult?.blake256 }}
     </div>
-    <div class="m-full">
-      <label class="block text-left pt-2 text-gray-800 text-sm font-bold mb-1" for="username">
-        Target Path
-      </label>
-      <div class="bg-slate-400 shadow-md rounded px-3 pt-4 pb-1 ">
-        <div class="mb-2">
-          <o-field>
-            <o-input placeholder="..." :model-value="filePath" type="text" expanded> </o-input>
-          </o-field>
-          <button @click="openFileDialog()" class="bg-blue-500 rounded-md text-white text-lg font-bold mt-2 w-full">Load
-            File</button>
-        </div>
-      </div>
-
-      <div class="overflow-scroll" style="height: 250px;">
-        <label class="block text-left text-gray-600 font-bold mb-1 pt-2" for="blake2b">Blake2s256</label>
-        <div
-          class="bg-slate-700 break-words text-white max-h-56 overflow-auto fullshadow-md rounded px-2 pt-1 pb-1  text-sm">
-          {{ hash_result_256 }}
-        </div>
-        <label class="block text-left text-gray-600 font-bold mb-1 pt-2" for="blake2b">Blake2b512</label>
-        <div
-          class="bg-slate-700 break-words text-white max-h-56 overflow-auto fullshadow-md rounded px-2 pt-1 pb-1  text-sm">
-          {{ hash_result_512 }}
-        </div>
-      </div>
+    <label class="block text-left text-gray-600 font-bold mb-1 pt-2" for="blake2b">Blake2b512</label>
+    <div
+      class="bg-slate-700 break-words text-white max-h-56 overflow-auto fullshadow-md rounded px-2 pt-1 pb-1  text-sm">
+      {{ hashResult?.blake512 }}
     </div>
   </div>
 </template>
