+++ src/src/types/FileProps.ts
@@ -1,0 +1,6 @@
+export type FileProps = {
+    filePath: string,
+    fileName: string,
+    fileSize: string,
+    fileExtension: string
+}
