+++ src/src/types/HashResult.ts
@@ -1,0 +1,4 @@
+export type HashResult = {
+    blake512: string,
+    blake256: string
+}
